nexus-ai/requirements.txt
fastapi
uvicorn
scikit-learn
scipy
pydantic
python-dotenv
requests 
google-cloud-aiplatform 